+++ CHANGES.md
@@ -11,6 +11,7 @@ Improvements 🙌:
  - Small optimisation of scrolling experience in timeline (#2114)
  - Allow user to reset cross signing if he has no way to recover (#2052)
  - Create home shortcut for any room (#1525)
+ - Filter room member (and banned users) by name (#2184)
 
 Bugfix 🐛:
  - Improve support for image/audio/video/file selection with intent changes (#1376)